Refactored util files
