+++ lib/measure_repository_service_test_kit/measure_group.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
 require 'json'
-require_relative '../utils/package_utils'
+require_relative '../utils/general_utils'
 
 module MeasureRepositoryServiceTestKit
   # tests for read by ID and search for Measure service
   # rubocop:disable Metrics/ClassLength
   class MeasureGroup < Inferno::TestGroup
-    include PackageUtils
+    include GeneralUtils
 
     title 'Measure Read by Id and Search'
     description 'Ensure measure repository service can retrieve Measure resources by the server-defined id and search'
